requirements.txt
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==1.11.10
django-ckeditor==5.4.0
django-import-export==0.7.0
django-js-asset==1.0.0
djangorestframework==3.7.7
editdistance==0.4
et-xmlfile==1.0.1
gunicorn==19.6.0
idna==2.6
jdcal==1.3
numpy==1.14.0
odfpy==1.3.6
openpyxl==2.5.0
pandas==0.22.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
whitenoise==3.3.1
